Extract to extension for SiteTree related documents
